+++ runway/__init__.py
@@ -1,2 +1,2 @@
 """Set package version."""
-__version__ = '0.33.0'
+__version__ = '0.34.0'